Move the invalid encoding UI from `project` to `workspace`
