+++ crates/fs/src/fs.rs
@@ -7,7 +7,6 @@ pub mod fs_watcher;
 use anyhow::{Context as _, Result, anyhow};
 #[cfg(any(target_os = "linux", target_os = "freebsd"))]
 use ashpd::desktop::trash;
-use encodings::Encoding;
 use futures::stream::iter;
 use gpui::App;
 use gpui::BackgroundExecutor;
